refactor: shuffle unit tests
