+++ unit_tests/sources/declarative/auth/test_token_auth.py
@@ -249,52 +249,41 @@ def test_api_key_authenticator_inject(
     )
     assert {expected_field_name: expected_field_value} == getattr(token_auth, validation_fn)()
 
-def test_api_key_authenticator_nested_token_injection():
+
+@pytest.mark.parametrize(
+        "field_path, token, expected_result",
+        [
+            (
+                ["data", "auth", "token"],
+                "test-token",
+                {"data": {"auth": {"token": "test-token"}}},
+            ),
+            (
+                ["api", "{{ config.api_version }}", "auth", "token"],
+                "test-token",
+                {"api": {"v2": {"auth": {"token": "test-token"}}}},
+            ),
+        ],
+        ids=["Basic nested structure", "Nested with config interpolation"]
+)
+def test_api_key_authenticator_nested_token_injection(field_path, token, expected_result):
     """Test that the ApiKeyAuthenticator can properly inject tokens into nested structures when using body_json"""
     config = {"api_version": "v2"}
     parameters = {"auth_type": "bearer"}
     
-    test_cases = [
-        # Basic nested structure
-        {
-            "field_path": ["data", "auth", "token"],
-            "token": "test-token",
-            "expected": {"data": {"auth": {"token": "test-token"}}}
-        },
-        # Nested with config interpolation
-        {
-            "field_path": ["api", "{{ config.api_version }}", "auth", "token"],
-            "token": "test-token",
-            "expected": {"api": {"v2": {"auth": {"token": "test-token"}}}}
-        },
-        # Nested with parameter interpolation
-        {
-            "field_path": ["auth", "{{ parameters.auth_type }}", "credentials"],
-            "token": "test-token",
-            "expected": {"auth": {"bearer": {"credentials": "test-token"}}}
-        },
-        # Deep nesting with both interpolations
-        {
-            "field_path": ["api", "{{ config.api_version }}", "auth", "{{ parameters.auth_type }}", "token"],
-            "token": "test-token",
-            "expected": {"api": {"v2": {"auth": {"bearer": {"token": "test-token"}}}}}
-        }
-    ]
-
-    for case in test_cases:
-        token_provider = InterpolatedStringTokenProvider(
-            config=config,
-            api_token=case["token"],
+    token_provider = InterpolatedStringTokenProvider(
+        config=config,
+        api_token=token,
+        parameters=parameters
+    )
+    token_auth = ApiKeyAuthenticator(
+        request_option=RequestOption(
+            inject_into=RequestOptionType.body_json,
+            field_path=field_path,
             parameters=parameters
-        )
-        token_auth = ApiKeyAuthenticator(
-            request_option=RequestOption(
-                inject_into=RequestOptionType.body_json,
-                field_path=case["field_path"],
-                parameters=parameters
-            ),
-            token_provider=token_provider,
-            config=config,
-            parameters=parameters,
-        )
-        assert token_auth.get_request_body_json() == case["expected"]
+        ),
+        token_provider=token_provider,
+        config=config,
+        parameters=parameters,
+    )
+    assert token_auth.get_request_body_json() == expected_result
